+++ x/zoneconcierge/keeper/epochs.go
@@ -3,40 +3,23 @@ package keeper
 import (
 	"context"
 
-	"cosmossdk.io/store/prefix"
 	epochingtypes "github.com/babylonlabs-io/babylon/v3/x/epoching/types"
 	"github.com/babylonlabs-io/babylon/v3/x/zoneconcierge/types"
-	"github.com/cosmos/cosmos-sdk/runtime"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // GetLastSentSegment get last broadcasted btc light client segment
 func (k Keeper) GetLastSentSegment(ctx context.Context) *types.BTCChainSegment {
-	store := k.storeService.OpenKVStore(ctx)
-	has, err := store.Has(types.LastSentBTCSegmentKey)
+	segment, err := k.LastSentBTCSegment.Get(ctx)
 	if err != nil {
-		panic(err)
-	}
-	if !has {
 		return nil
 	}
-	segmentBytes, err := store.Get(types.LastSentBTCSegmentKey)
-	if err != nil {
-		panic(err)
-	}
-	var segment types.BTCChainSegment
-	k.cdc.MustUnmarshal(segmentBytes, &segment)
 	return &segment
 }
 
 // setLastSentSegment sets the last segment which was broadcasted to the other light clients
 // called upon each AfterRawCheckpointFinalized hook invocation
-func (k Keeper) setLastSentSegment(ctx context.Context, segment *types.BTCChainSegment) {
-	store := k.storeService.OpenKVStore(ctx)
-	segmentBytes := k.cdc.MustMarshal(segment)
-	if err := store.Set(types.LastSentBTCSegmentKey, segmentBytes); err != nil {
-		panic(err)
-	}
+func (k Keeper) setLastSentSegment(ctx context.Context, segment *types.BTCChainSegment) error {
+	return k.LastSentBTCSegment.Set(ctx, *segment)
 }
 
 func (k Keeper) GetLastFinalizedEpoch(ctx context.Context) uint64 {
@@ -54,26 +37,15 @@ func (k Keeper) recordSealedEpochProof(ctx context.Context, epochNum uint64) {
 		panic(err) // only programming error
 	}
 
-	store := k.sealedEpochProofStore(ctx)
-	store.Set(sdk.Uint64ToBigEndian(epochNum), k.cdc.MustMarshal(proofEpochSealed))
+	if err := k.SealedEpochProof.Set(ctx, epochNum, *proofEpochSealed); err != nil {
+		panic(err)
+	}
 }
 
 func (k Keeper) getSealedEpochProof(ctx context.Context, epochNum uint64) *types.ProofEpochSealed {
-	store := k.sealedEpochProofStore(ctx)
-	proofBytes := store.Get(sdk.Uint64ToBigEndian(epochNum))
-	if len(proofBytes) == 0 {
+	proof, err := k.SealedEpochProof.Get(ctx, epochNum)
+	if err != nil {
 		return nil
 	}
-	var proof types.ProofEpochSealed
-	k.cdc.MustUnmarshal(proofBytes, &proof)
 	return &proof
-}
-
-// sealedEpochProofStore stores the proof that each epoch is sealed
-// prefix: SealedEpochProofKey
-// key: epochNumber
-// value: ProofEpochSealed
-func (k Keeper) sealedEpochProofStore(ctx context.Context) prefix.Store {
-	storeAdapter := runtime.KVStoreAdapter(k.storeService.OpenKVStore(ctx))
-	return prefix.NewStore(storeAdapter, types.SealedEpochProofKey)
 }